changed size_t to int

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -10,7 +10,7 @@
 size_t list_len(const list_t *h)
 {
 	const list_t *currentNode;
-	size_t nodes;
+	int nodes;
 
 	nodes = 0;
 	currentNode = malloc(sizeof(list_t));
